Kernel v4.9.5 fix and module reload fix
For the v4.9.5 fix, we disable MSI initially to avoid a crash
when the nvme driver allocates an initial MSI vector. We re-enable
MSI later so that the interrupt vectors can be allocated.

For the module reload fix, pci_remove_bus wasn't fully clearing
away directories, and virt couldn't start on a reload.
pci_remove_root_bus seems to clear everything. The second part
of the module reload fix is sending the early legacy interrupts
to the interrupt's affinity vector instead of CPU 0. On boot,
virt's interrupt 16 (IRQ_NUM) seemed to always be assigned to
CPU 0 and others. However, on module unload and reload, this
wasn't always the case. We just send the interrupts to the
whole vector and one will pick them up. Produces a few IRQ
handler warnings which can maybe be looked at later, but for
now don't seem to be important.